Refactor fss monitor
Calculate used disk space with java code instead of using du command

LMCROSSITXSADEPLOY-479

+++ com.sap.cloud.lm.sl.cf.web/src/main/java/com/sap/cloud/lm/sl/cf/web/monitoring/Metrics.java
@@ -45,16 +45,16 @@ public class Metrics implements MetricsMBean {
     }
 
     @Override
-    public double getUsedFssSpace() {
+    public long getUsedFssSpace() {
         if (!shouldCollectCentralServiceMetrics() || getFssStoragePath().isEmpty()) {
             LOGGER.debug("Not collecting metrics for FSS on path: {}", getFssStoragePath());
-            return 0d;
+            return 0L;
         }
         return fssMonitor.calculateUsedSpace(fileSystemService.getStoragePath());
     }
 
     @Override
-    public double getUsedContainerSpace() {
+    public long getUsedContainerSpace() {
         String workDir = System.getProperty("user.dir");
         String parentDir = Paths.get(workDir)
             .getParent()
